release 0.1.71 source code

+++ services/dgc/v1/model/model_node_instance.go
@@ -11,13 +11,13 @@ type NodeInstance struct {
 
 	Status *string `json:"status,omitempty"`
 
-	PlanTime *int32 `json:"planTime,omitempty"`
+	PlanTime *int64 `json:"planTime,omitempty"`
 
-	StartTime *int32 `json:"startTime,omitempty"`
+	StartTime *int64 `json:"startTime,omitempty"`
 
-	EndTime *int32 `json:"endTime,omitempty"`
+	EndTime *int64 `json:"endTime,omitempty"`
 
-	ExecuteTime *int32 `json:"executeTime,omitempty"`
+	ExecuteTime *int64 `json:"executeTime,omitempty"`
 
 	NodeType *string `json:"nodeType,omitempty"`
